requirements.txt
crossrefapi
selenium
webdriver-manager
psutil